Convert motion text from hiragana to kanji
The next obvious step.  I've used google's transliteration service to do
this for now, with a hardcoded call to system curl, and making use of Po
Shan Cheah's "ParseJSON" code.  I plan to separate this out into a
separate file to make it easy to write IME "plugins" to interface with
different transliteration engines (SCIM, ibus, whatever), but for now
this'll do to get things up and running.