Add highlighting to the word being typed
This is a pretty poor implementation at the moment but right now I
haven't got a better way.  Basically it looks at the word between the
kanji start point and the current control and sets a syntax highlight on
it.  Rubbish, because if that word (and to start with, it's just the
first letter!) appears anywhere else in the document it'll be
highlighted everywhere.