Activate vimes before motion-driven kanji conversion
Doing this means all the vimes keymappings, etc will be activated.
We have to disable the insert mode mapping on our way in, as we are
setting the value of g:kanji_start_point outside of insert mode and we
don't want it to be reset, however it needs to be re-enabled once we're
in, in case the user leaves insert mode without de-activating vimes.